[IMP] sale, account, base_setup, purchase, base, report_intrastat:  Address Justification mode of destinator's address on printed documents.
Currently the position for customer address on printable documents is fixed
while it may vary from one country to another (and so is it for envelope windows)
like in Europe: center to right side of the page
and in US: it is most common on the left side, justified with the rest of the text.

so in this task we added a setting in general configuration settings:
Address Justification / Justification mode of destinator's address on printed documents
- with radio button: Left/right we can change customer address position
on report. (customer chose left alignment we should swap the address
 locations and put invoicing & billing address on right side)

- when multicompany enabled and user has different company then document,
then print Address justification position based on document assigend
company not user allocated company.

This commit is related to task ID #34078
Closes #18683

[FIX]