[ADD] stock: Report Forecasted Inventory
Purpose of this task:
Show forecasted inventory for all products.
Before this commit we have to search forecasted inventory for each product individually.
There is no way to check all the forecasted inventory in a single menu.

After this commit:
User can search for all forecasted inventory from the menu 'Forecasted Inventory'.
Features added.
1. Menu for Forecasted Inventory has been added.
2. List and pivot view for Forecasted Inventory added.
3. Default measures has been set to 'Quantity' and 'Cumulative Quantity'.
4. Pivot view is grouped by product (vertical) and by weeks (horizontal).

This commit is related to task 36088.