[ADD] stock: Query changes for report Forecasted Inventory Purpose of this task: Show forecasted inventory for all products. Before this commit we have to search forecasted inventory for each product individually. There is no way to check all the forecasted inventory in a single menu.
After this commit:
User can search for all forecasted inventory from the menu 'Forecasted Inventory'.
Features added.
1. lines only those containing any stock move will be shown.
2. Reference field added.
We remove 'week' from the query, instead of we fetch all the data according to 'date_expected', and we create three group_bys (i.e. group by date, week and month) so it is easy for the user to check report according to day, week and month.

This commit is related to task 36088.