[IMP] product, website_sale :  Improve behaviour of product images for e-Commerce
Previously, we used to display main product image(of a variant or template) and the
additional images(from product template) on eCommerce product page.
Also, setting multiple images per variant was not possible.

It was causing inconsistency in size of the images. When user changes the main image,
it is always resized by the framework (so even if user uploads larger image, it is
squared up to max 1024 x 1024 px). While it is not the case with the additional
images(product_image_ids) because original size is preserved. This behaviour was
confusing for users because they expect to see the all the images they uploaded in
original size on eCommerce’s product page, including the main image.

We can’t change the mechanism of main product image being resized, as resized images
are expected in other modules, like POS.

So as a solution, this commit implies below mentioned changes :

- Introduced new sequence field on product.image table
- In product template, there will always be a copy of main image in additional images,
  which will have sequence=0 (to identify it as a main image, and update it automatically
  when user changes the main image) and will also preserve original size.
- On eCommerce page, this copy of image with original size will be displayed instead of the
  main image

In addition to this there are several other improvements
- Now user will be able to add multiple images per variant also. Mechanism for resizing and
  keeping copy of original image to preserve size remains same then product template.
- On product carousel, variant specific images will be displayed first, and then common template
  images, both sorted by sequence.
- In case of more images, instead of displaying scroll bar, indicators are added to carousel
- To keep edition feature enabled from front-end, we render carousel template from serverside

Related Task : 34045
Closes : #24631