[ADD] web: add edit tag feature on many2many_tags widget, also provide no_edit option to remove edit feature from many2many_tags widget
open a customer form from a meeting (to see the address, the phone, etc)
In case of mispelling, etc you need to go to the settings to change a m2m record. Most of the times menu items are hidden so its painful: product attributes, customer tags
the issue is worse in settings: in new settings there is a lot of m2m tag you cannot change from this screen.

with this commit clicking on m2m tag will give Edit option unless no_edit option is explicitly defined in xml view on field, edit form dialog is given to user to edit m2m tag

Related to task #29227
PR  #22957

Co-authored-by: Mohammed Shekha <msh@openerp.com>