[FIX] website_sale: Prevent updation of cart qty in edit mode
Because of the mechanism to bind change / click event in current implementation,
when we are in edit mode of a product page, respected events are triggered on
change of cart qty input (responsible class : js_quantity) or on click of
+ / - buttons (responsible class : js_add_cart_json). And this is responsible
of triggering further events (calling update_product_image method, which
causes this issue).

After this commit, click / change event while updating cart qty will not be
effective in edit mode, preventing further unwanted events being triggered.

Related Task : 1825889
PR : #24211