[REF] mail: use a query to fetch notification recipients data
In this commit we update _notify method of mail.message to fetch most
data related to notification process in a single query.  Indeed posting a
message can be a costly process, especially when involving several partners,
channels, depending on their notification status, user groups, ... Most
of the required data can be fetched in a single hand-tuned query. Purpose
is to lessen query count. Side effect of this commit is that code is less
recordset based and therefore a bit more low-level.

API of notification methods is updated to allow passing notably the values
used to create the newly-posted message. It allows in some simple case to
avoid browsing message and partners when no notification is necessary. It
saves some queries in simple case like posting messages with specific subtypes
not often followed or logging notes without mentioning people.

Functionally this commit should not imply any change. This commit allows to
save a few queries on several performance tests as indicated by updated
performance counters. This commit is linked to task ID 47934.