[IMP] test_mail: add base tests about tracking
Purpose of this commit is to add small performance tests for some base
features in tracking. Two base use case of tracking are added in this
commit:

 * base tracking of 2 values, without any subtype;
 * tracking of 2 values triggering a message post with a subtype and without
   specific followers;

Having followers is already covered by other tests as it will simply increase
query count depending on notification choice (inbox / email).

In current saas-11.2 cost of tracking is mainly linked to cost of logging
a note or posting a message with a subtype as creating tracking values is
done as sudo and adds few queries.