[REF] mail: clean code related to notification email generation
This commit cleans code about notification in partner. Its purpose is to
prepare notification emails, group recipients and finally send emails. It
is done by

 * propagating some additional parameter to simplify code, like record
   on which the notification process runs;
 * inlining some code currently located in sub-methods;
 * simplifying a bit notification emails preparation and computation;

It allows to save a few query by avoiding to fetch and browse again some
data that was already computed.

This commit is linked to task ID 47934. No functional change should occur
with this commit.