[REF] mail: propagate recipient data through notification process
Purpose of this commit is to avoid browsing and prefetching data about
recipients to notify of a new message. As we now have a hand-tuned query
that fetches all required data let us propagate its use through various
methods instead of (pre)fetching / browsing data we already have.

This commit is linked to task ID 47934. No functional change should occur
with this commit.