[FIX] web: discarding a popup with ESCAPE key do not create an error
Before this fix, when opening a form that has a many2one in a popup,
itself containing a form with a one2many in create mode, when discarding
the second popup with the ESCAPE key, there was an error.

After this fix, we catch the event triggered when the ESCAPE key in the
form_dialog, cancel the event and close the dialog. This prevents the
error and gives a better behavior.