[REF] test_mail: update counters due to enterprise mail_push update
Linked to: [REF] mail_push: rewrite part of notification override

Enterprise commit XXXXXX updated code of mail_push to lessen its impact on
mail-related code. It allows to change a few counters in test_mail that is
impacted by mail_push as this module is auto install.

This commit is linked to task ID 47934. No functional change should occur
with this commit.