[IMP] web: one to many keyboard navigation
Before this commit, when adding a new line in a one2many editable list,
if the user left the first column empty and hit TAB, the row was
discarded and the focus moved to the next widget of the form.

After this commit, when hitting TAB will not discard the row and will
move the focus to the  next column. Hitting ESCAPE will discard the
row and move the focus to the  "add line"  link or button.