[IMP] account: improved reconcilation tour and qunit test case.
On this task currency position and the presence of a space between the amount and symbol are defined on the language. so default currency selected on reconcilation tour and test cases are based on language which contain no space between Currency symbol and amount and that's why We need to remove space between amount and symbol on tour and  on reconcile qunit test case too which contain space between amount and currency symbol on older test case.

Related TaskId: 35636. Closes: #25183.