[IMP] base : Direct access to recors rules and access rights from user
Add 2 smart buttons with count to see :
- Access Controls
- Record Rules
Buttons available in debug mode only
Groups smart button not added as duplicate of Group tab.
Every user that have access to user form has at least admin / Settings or Admin / Access rights.
Admin / Settings includes Admin / Access Rights
As showing the button only on debug mode, it's useless to specifiy
the group erp_manager for those fields.

Create and delete are not allowed for those two views as it's could let the user think
that he will add an access right only for the active_id
and access rights and record rules are managed per group, not per user.
To avoid confusion, the user's access rights list is now not editable.
Clicking on a row redirect to form view that can be edited.

Co-authored-by: David Beguin <dbe@odoo.com>
Co-authored-by: FMDL <florent.mirieu@gmail.com>