[IMP] mail: Allowing drag/drop files and paste image on mail composer
By this commit user can upload attachments by drag and drop in chatter composer.
User can also attach image in composer by paste. This feature can be used from
discuss UI, chat windows and form view chatter.

Before this commit chatter unlinks the attachment if another attachment uploaded
with same name. Criteria for removing duplicate attachment is changed with this
commit. Now it will remove the attachment based on name and size of attachment.
The reason for this change is, when you paste image from clipboard it always have
same name so with old condition it was not possible to attach more then one image
by paste.

Related to task #32469
Closes #18914