[IMP] payment: support SEPA Direct Debit payment acquirer
Added client side BIC validation on checkout form for improved UX.

Opening an invoice related to a SEPA Direct Debit mandate was causing the invoice to be paid and therefore the related payment to be posted.
A new payment was then created based on the transaction and was then posted. We ended up with 2 payments being posted.
Added some checks on the invoice and payment status to avoid this issue.

Related to task #1870363
Closes #26427