[IMP] models: add filter on access rules
Currently check_access_rule method checks for ir rules and raises an error
if there are either missing ids (deleted records) or forbidden ids (records
hidden due to an ir rule).

Purpose of this commit is to have a method allowing to filter a given record
set and receive only valid records. It implementation is basically the same
as old check_access_rule with an optional raise flag that is used by
check_access_rule to avoid code duplication.

This commit is linked to task ID 1856417 and PR #26565.