[FIX] hr_timesheet: moving a task from project
When moving a task, we want the timesheets of the task to
follow the move. (This will not be the case in saas11.5).
However, a error is raised when moving a task in a project
not linked to a AA (required to timesheet on it). This error
makes sens when the moving task has timesheets. If not, the
error is still raise, and makes no sense.
This commit ensures the error is raised only if the moving
task has some timesheets.