[FIX] stock: force save location_dest_id
When modifying a stock.picking, we might want to change the partner. This change
triggers a change on the destination through onchange. But since the field is
readonly, when saving the record, the field is not send to the server and so
the new value of destination is not saved on the picking, neither propagate
to the stock.moves.
This first part brings problems: it will be possible to change the
location of a picking but not the one from its move lines, when the
picking is confirmed. This can cause many many inconsistencies. The
solution is to prevent changing the partner (and not trigger the
onchange) when the state is not 'draft'.

opw-1872635