(WIP)[FIX] web: use read to update record values after resequence
Commit:
https://github.com/odoo/odoo/commit/818c18e55d0718286ff5bf332186a10f4d7a58ef
Updated the resequence logic in a way that was almost falser than before.
The added test however did work by coincidence, as index values were equal to
the sequence field values.
To be sure that we synchronize with what happens on the server, we do a read
after the resequence.
This as the added benefit of breaking kanban resequencing behaviour: if the
sequence field does not exist, the kanban objects are reorganized anyway, even
if they are not really resequenced.
Since now we check with a read, this is not true anymore.
It's a WIP, so what happens then has not been decided yet.

opw 1867049