[FIX] account,web: fix invoice iframe height
a7a89f31741c3df1d15f1fb8d0433b2761a7e1d0 introduced a change that broke the invoice iframe.

The #wrapwrap corresponds to the client body, therefore the container must be inside it, and the scrollbar will now be on it instead of on the iframe body.

PR: #26801
task-1876864