[IMP] sale: improve demo, default data, and manifests
*** Activate pay and sign, as well as quotation template, by default when using demo data.

Previously, the features would show by default on runbot because they activated on installing modules.

Now that they are settings of sale_management, we need to specifically activate them for demoing. However we don't want them for on-boarding.

*** Remove the default template from sale_management, move it into sale_quotation_builder

The default template was useless in sale_management because it was empty: it was previously a required field, which is not the case anymore.

It was only filled and used in sale_quotation_builder to display an example quotation, which is why it is now moved there.

*** Update manifests

Fix category of sale_quotation_builder and slightly improve its text.

PR: #26801
task-1876864